requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
urllib3==2.0.4
python-dotenv==1.0.0
python-multipart==0.0.20
opencv-python==4.11.0.86
numpy==2.1.1
RPi.GPIO==0.7.1
Pillow==10.1.0
ultralytics==8.0.196
torch>=1.8.0
torchvision>=0.9.0